Add static lock cluster
