+++ picotest_macros/src/utils.rs
@@ -1,37 +1,33 @@
-use std::collections::VecDeque;
-use syn::UseTree;
+use syn::{parse_quote, Attribute, FnArg, ItemFn, Stmt};
+const TEST_PREFIX: &str = "test_";
 
-fn recurse_use_item<'a>(use_tree: &'a UseTree, tokens: &mut VecDeque<&str>) -> Option<&'a UseTree> {
-    let curr_ident = tokens.pop_front().unwrap_or_default();
-    match use_tree {
-        UseTree::Path(use_stmt_third) => {
-            if use_stmt_third.ident == curr_ident {
-                recurse_use_item(&use_stmt_third.tree, tokens)
-            } else {
-                None
-            }
-        }
-        UseTree::Name(ref use_stmt_third) => {
-            if use_stmt_third.ident == curr_ident {
-                Some(use_tree)
-            } else {
-                None
-            }
-        }
-        UseTree::Group(ref use_stmt_third) => {
-            for item in &use_stmt_third.items {
-                tokens.push_front(curr_ident);
-                if let Some(group_tree) = recurse_use_item(item, tokens) {
-                    return Some(group_tree);
-                }
-            }
-            None
-        }
-        _ => None,
+pub fn process_test_function(mut func: ItemFn, path: &String, timeout: u64) -> ItemFn {
+    let func_name = func.sig.ident.to_string();
+    if !func_name.starts_with(TEST_PREFIX) {
+        return func;
     }
-}
 
-pub fn traverse_use_item<'a>(use_tree: &'a UseTree, tokens: Vec<&str>) -> Option<&'a UseTree> {
-    let mut tokens = VecDeque::from(tokens);
-    recurse_use_item(use_tree, &mut tokens)
+    let rstest_macro: Attribute = parse_quote! { #[rstest] };
+    func.attrs.insert(0, rstest_macro);
+
+    let cluster: FnArg = parse_quote! {
+        #[with(#path, #timeout)] cluster: &Cluster
+    };
+    func.sig.inputs.insert(0, cluster);
+
+    let block = func.block.clone();
+    let new_body: Stmt = parse_quote! {
+        let result = panic::catch_unwind(panic::AssertUnwindSafe(|| {
+            #block
+        }));
+    };
+
+    let resume: Stmt = parse_quote! {
+        if let Err(err) = result {
+            panic::resume_unwind(err);
+        }
+    };
+    func.block.stmts = vec![new_body, resume];
+
+    func
 }
